% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_export}
\alias{df_export}
\title{Transfer row.names values to a new column}
\usage{
df_export(input_table, new_rn = "variable")
}
\arguments{
\item{rich}{A dataframe with alpha diversity indices as columns
and samples as rows as the output of pyloseq::estimate_richness()}

\item{file}{Output file name}
}
\value{
Input data.frame with an extra column contains
}
\description{
Transfer row.names values to a new column
}
