% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{ggrare}
\alias{ggrare}
\title{Plot rarefaction curves}
\usage{
ggrare(
  physeq,
  step = 10,
  label = NULL,
  color = NULL,
  plot = TRUE,
  parallel = FALSE,
  se = TRUE
)
}
\arguments{
\item{physeq}{A phyloseq class object, from which abundance
data are extracted}

\item{step}{Step size for sample size in rarefaction curves}

\item{label}{Default \code{NULL}. Character string.
The name of the variable to map to text labels on the plot.
Similar to color option but for plotting text.}

\item{color}{(Optional). Default ‘NULL’. Character string.
The name of the variable to map to colors in the plot.
This can be a sample variable (among the set returned by
‘sample_variables(physeq)’ ) or taxonomic rank (among the set returned by
‘rank_names(physeq)’).
The color scheme is chosen automatically by
‘link{ggplot}’, but it can be modified afterward with an
additional layer using ‘scale_color_manual’.}

\item{plot}{Logical, should the graphic be plotted.}

\item{parallel}{should rarefaction be parallelized (using parallel framework)}

\item{se}{Default TRUE. Logical. Should standard errors be computed.}
}
\value{
Print ggplot rarefaction curves
}
\description{
The function plot rarefaction curves...
}
