% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NbClust_indexlist.R
\name{nb_clust_all}
\alias{nb_clust_all}
\title{Use NbClust::NbClust iteratively deal with errors}
\usage{
nb_clust_all(data, seed)
}
\arguments{
\item{data}{A dataframe with variables in column}

\item{seed}{An integer defining the seed}
}
\value{
Number of clusters per index
}
\description{
Use NbClust::NbClust iteratively deal with errors
}
