% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{primer_trim}
\alias{primer_trim}
\title{Primer trimming using cutadapt}
\usage{
primer_trim(
  forward_files,
  reverse_files,
  primer_fwd,
  primer_rev,
  output_dir,
  min_size
)
}
\arguments{
\item{forward_files}{vector of names of forward reads files to trim}

\item{reverse_files}{vector of names of reverse reads files to trim}

\item{primer_fwd}{forward primer}

\item{primer_rev}{reverse primer}

\item{output_dir}{folder name where to write trimmed sequences}

\item{min_size}{sequences shorter than min_size after trimming are
filtered out}
}
\value{
pair of gunzipped fastq files (R1 and R2) of primer trimmed sequences
}
\description{
Primer trimming using cutadapt
}
\details{
Input file names have to end with _R1.fastq.gz or _R2.fastq.gz.

Cutadapt need to be installed on your system.
}
