% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{qualityprofile}
\alias{qualityprofile}
\title{Plot one quality profile per fastq file}
\usage{
qualityprofile(filefw, filerev, outfile)
}
\arguments{
\item{filefw}{R2 file names}

\item{outfile}{Output pdf file name}
}
\value{
a pdf file containing one quality profile per file
}
\description{
Plot one quality profile per fastq file
}
