% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{run_cutadapt_command}
\alias{run_cutadapt_command}
\title{A function rather aimed at developers}
\usage{
run_cutadapt_command(
  forward_file,
  reverse_file,
  primer_fwd,
  primer_rev,
  output_dir,
  min_size,
  log_dir
)
}
\description{
A function that does blabla, blabla.
}
\keyword{internal}
